+++ src/app/calendar/ParcelCalendar.tsx
@@ -4,9 +4,9 @@ import Calendar from "@/components/Calendar/Calendar";
 import Title from "@/components/Title/Title";
 import React from "react";
 import styled, { RainbowPalette, useTheme } from "styled-components";
-import { ParcelWithClientName } from "@/app/calendar/getParcelsWithCollectionDate";
 import parcelsToCollectionEvents, {
     LocationColorMap,
+    ParcelWithClientName,
 } from "@/app/calendar/parcelsToCollectionEvents";
 
 interface ParcelCalendarProps {
